backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
pandas
reportlab
sqlalchemy
python-dotenv
aiohttp
